+++ pythran/pythonic/include/__builtin__/id.hpp
@@ -1,9 +1,8 @@
 #ifndef PYTHONIC_INCLUDE_BUILTIN_ID_HPP
 #define PYTHONIC_INCLUDE_BUILTIN_ID_HPP
 
-#include <cstdint>
-#include "pythonic/types/long.hpp"
-#include "pythonic/utils/proxy.hpp"
+#include "pythonic/include/types/long.hpp"
+#include "pythonic/include/utils/proxy.hpp"
 
 namespace pythonic
 {
@@ -12,12 +11,12 @@ namespace pythonic
   {
 
     template <class T>
-    intptr_t id(T const &t);
+    long id(T const &t);
 
-    intptr_t id(pythran_long_t const &t);
-    intptr_t id(long const &t);
-    intptr_t id(double const &t);
-    intptr_t id(bool const &t);
+    long id(pythran_long_t const &t);
+    long id(long const &t);
+    long id(double const &t);
+    long id(bool const &t);
 
     PROXY_DECL(pythonic::__builtin__, id);
   }
